# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FSTS Client Dashboard Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List all client sites visible to the current user
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create a new client site (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{siteId}:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getSite
      tags: [sites]
      summary: Get a single site
      responses:
        "200":
          description: Site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Update site metadata / branding / white-label settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Archive / delete a client site (super admin only, requires confirmation client-side)
      responses:
        "204":
          description: Deleted
  /sites/{siteId}/dashboard-summary:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getSiteDashboardSummary
      tags: [sites]
      summary: Get aggregate counts and recent activity for a site's dashboard home
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
  /sites/{siteId}/homepage:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getHomepageContent
      tags: [content]
      summary: Get the homepage content block for a site
      responses:
        "200":
          description: Homepage content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageContent"
    put:
      operationId: updateHomepageContent
      tags: [content]
      summary: Update the homepage content block for a site (creates a version snapshot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HomepageContentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageContent"
...
  /sites/{siteId}/events:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listEvents
      tags: [content]
      summary: List events for a site
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [content]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /sites/{siteId}/events/{eventId}:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getEvent
      tags: [content]
      summary: Get a single event
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [content]
      summary: Update an event (creates a version snapshot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [content]
      summary: Delete an event (requires confirmation client-side)
      responses:
        "204":
          description: Deleted
...
  /sites/{siteId}/footer:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getFooterContent
      tags: [footer-contact]
      summary: Get footer content for a site
      responses:
        "200":
          description: Footer content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FooterContent"
    put:
      operationId: updateFooterContent
      tags: [footer-contact]
      summary: Update footer content for a site (creates a version snapshot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FooterContentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FooterContent"
...
components:
  schemas:
...
    Site:
      type: object
      required: [id, name, slug, status, brandColorPrimary, brandColorSecondary, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        status:
          type: string
          enum: [active, staging, archived]
        domain: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        faviconUrl: { type: ["string", "null"] }
        brandColorPrimary: { type: string }
        brandColorSecondary: { type: string }
        whiteLabelEnabled: { type: boolean }
        poweredByFsts: { type: boolean }
        createdAt: { type: string, format: date-time }
    SiteInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        status:
          type: string
          enum: [active, staging, archived]
        domain: { type: string }
        logoUrl: { type: string }
        faviconUrl: { type: string }
        brandColorPrimary: { type: string }
        brandColorSecondary: { type: string }
        whiteLabelEnabled: { type: boolean }
        poweredByFsts: { type: boolean }
    SiteUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        status:
          type: string
          enum: [active, staging, archived]
        domain: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        faviconUrl: { type: ["string", "null"] }
        brandColorPrimary: { type: string }
        brandColorSecondary: { type: string }
        whiteLabelEnabled: { type: boolean }
        poweredByFsts: { type: boolean }
    DashboardSummary:
      type: object
      required: [siteId, courseCount, eventCount, articleCount, mediaCount, recentActivity]
      properties:
        siteId: { type: integer }
        courseCount: { type: integer }
        eventCount: { type: integer }
        articleCount: { type: integer }
        mediaCount: { type: integer }
        lastBackupAt: { type: ["string", "null"], format: date-time }
        squareConnected: { type: boolean }
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLogEntry"
    HomepageContent:
      type: object
      required: [siteId, heroHeadline, heroSubheadline, sections, updatedAt]
      properties:
        siteId: { type: integer }
        heroHeadline: { type: string }
        heroSubheadline: { type: string }
        heroImageUrl: { type: ["string", "null"] }
        sections:
          type: array
          items:
            type: object
            additionalProperties: true
        updatedAt: { type: string, format: date-time }
    HomepageContentInput:
      type: object
      required: [heroHeadline, heroSubheadline]
      properties:
        heroHeadline: { type: string, minLength: 1 }
        heroSubheadline: { type: string, minLength: 1 }
        heroImageUrl: { type: string }
        sections:
          type: array
          items:
            type: object
            additionalProperties: true
...
    Event:
      type: object
      required: [id, siteId, title, slug, status, description, startAt, createdAt, updatedAt]
      properties:
        id: { type: integer }
        siteId: { type: integer }
        title: { type: string }
        slug: { type: string }
        status:
          type: string
          enum: [draft, published, archived]
        description: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: ["string", "null"], format: date-time }
        location: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EventInput:
      type: object
      required: [title, slug, description, startAt]
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        status:
          type: string
          enum: [draft, published, archived]
        description: { type: string, minLength: 1 }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        location: { type: string }
        imageUrl: { type: string }
    EventUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        status:
          type: string
          enum: [draft, published, archived]
        description: { type: string, minLength: 1 }
        startAt: { type: string, format: date-time }
        endAt: { type: ["string", "null"], format: date-time }
        location: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
...
    FooterContent:
      type: object
      required: [siteId, columns, socialLinks, copyrightText, updatedAt]
      properties:
        siteId: { type: integer }
        columns:
          type: array
          items:
            type: object
            additionalProperties: true
        socialLinks:
          type: array
          items:
            type: object
            additionalProperties: true
        copyrightText: { type: string }
        updatedAt: { type: string, format: date-time }
    FooterContentInput:
      type: object
      required: [copyrightText]
      properties:
        columns:
          type: array
          items:
            type: object
            additionalProperties: true
        socialLinks:
          type: array
          items:
            type: object
            additionalProperties: true
        copyrightText: { type: string, minLength: 1 }
...
    ActivityLogEntry:
      type: object
      required: [id, siteId, actorName, action, entityType, createdAt]
      properties:
        id: { type: integer }
        siteId: { type: integer }
        actorName: { type: string }
        action: { type: string }
        entityType: { type: ["string", "null"] }
        entityId: { type: ["integer", "null"] }
        details: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...